feat: single scaffold
Thank god!
To be honest, this doesn't look great yet, but it's technically correct.
This should make life significantly less buggy.

Note: there seems to be a bug when switching between desktop and mobile.
Is that really a common situation outside of development?
Ehh, hopefully not.
