+++ packages/app/lib/src/features/home/presentation/home/feed.dart
@@ -23,7 +23,7 @@ class Feed extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    // Maybe change to scaffold with floating action button and list view as child
+    // TODO(MattsAttack): Maybe change to scaffold with floating action button and list view as child.
     return Container(
       alignment: Alignment.center,
       constraints: const BoxConstraints(maxWidth: 600),
