+++ packages/app/lib/src/features/home/presentation/home/feed_routing_page.dart
@@ -6,6 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../../../app/router.gr.dart';
+import '../../../../app/wrapper.dart';
 
 /// {@template nexus.features.home.presentation.home.feed_routing_page}
 /// A page that manages routing between different feeds.
@@ -19,23 +20,21 @@ class FeedRoutingPage extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return AutoTabsScaffold(
+    final tabsRouter = context.tabsRouter;
+
+    return AutoTabsRouter.tabBar(
       routes: const [LocalFeedRoute(), WorldFeedRoute()],
-      appBarBuilder: (context, autoRouter) {
-        return AppBar(
-          scrolledUnderElevation: 0,
-          backgroundColor: Theme.of(context).colorScheme.surface,
-          automaticallyImplyLeading: false,
-          title: NavigationBar(
-            selectedIndex: autoRouter.activeIndex,
-            onDestinationSelected: autoRouter.setActiveIndex,
-            elevation: 0,
-            backgroundColor: Theme.of(context).colorScheme.surface,
-            destinations: const [
-              NavigationDestination(icon: Icon(Icons.pin_drop), label: 'Local'),
-              NavigationDestination(icon: Icon(Icons.public), label: 'World'),
+      builder: (context, child, tabController) {
+        return Wrapper(
+          autoRouter: tabsRouter,
+          tabs: TabBar(
+            controller: tabController,
+            tabs: const [
+              Tab(icon: Icon(Icons.pin_drop), text: 'Local'),
+              Tab(icon: Icon(Icons.public), text: 'World'),
             ],
           ),
+          child: child,
         );
       },
     );
