+++ packages/app/lib/src/features/map/presentation/items/map_info_page.dart
@@ -5,6 +5,7 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
+import '../../../../app/wrapper.dart';
 import '../../../home/domain/feed_entity.dart';
 import '../../../home/presentation/home/feed.dart';
 
@@ -30,7 +31,10 @@ class MapInfoPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Feed(feed: FeedEntity.local(latitude, longitude));
+    return Wrapper(
+      autoRouter: context.tabsRouter,
+      child: Feed(feed: FeedEntity.local(latitude, longitude)),
+    );
   }
 
   // coverage:ignore-start
