+++ packages/app/lib/src/features/settings/presentation/preferences/settings_page.dart
@@ -9,6 +9,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:url_launcher/link.dart';
 
 import '../../../../app/router.gr.dart';
+import '../../../../app/wrapper.dart';
 import '../../../../gen/assets.gen.dart';
 import '../../../../gen/version.gen.dart';
 import '../../../../utils/hooks.dart';
@@ -37,97 +38,100 @@ class SettingsPage extends ConsumerWidget {
     final settingsService = ref.watch(settingsServiceProvider);
     final themeMode = settingsService.themeMode;
 
-    return Padding(
-      padding: const EdgeInsets.all(16),
-      child: Column(
-        spacing: 8,
-        children: [
-          Padding(
-            padding: const EdgeInsets.all(16),
-            child: Container(
-              alignment: Alignment.topLeft,
-              child: DropdownMenu(
-                // Read the selected themeMode from the controller
-                initialSelection: themeMode,
-                // Call the updateThemeMode method any time the user selects a theme.
-                onSelected: (theme) async {
-                  final newTheme = theme ?? settingsService.themeMode;
+    return Wrapper(
+      autoRouter: context.tabsRouter,
+      child: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Column(
+          spacing: 8,
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(16),
+              child: Container(
+                alignment: Alignment.topLeft,
+                child: DropdownMenu(
+                  // Read the selected themeMode from the controller
+                  initialSelection: themeMode,
+                  // Call the updateThemeMode method any time the user selects a theme.
+                  onSelected: (theme) async {
+                    final newTheme = theme ?? settingsService.themeMode;
 
-                  await ref
-                      .read(settingsServiceProvider.notifier)
-                      .updateThemeMode(newTheme);
-                },
-                label: const Text('Theme'),
-                dropdownMenuEntries: const [
-                  DropdownMenuEntry(
-                    value: ThemeMode.system,
-                    label: 'System Theme',
-                    leadingIcon: Icon(Icons.brightness_medium),
-                  ),
-                  DropdownMenuEntry(
-                    value: ThemeMode.light,
-                    label: 'Light Theme',
-                    leadingIcon: Icon(Icons.brightness_5),
-                  ),
-                  DropdownMenuEntry(
-                    value: ThemeMode.dark,
-                    label: 'Dark Theme',
-                    leadingIcon: Icon(Icons.brightness_3),
-                  ),
-                ],
+                    await ref
+                        .read(settingsServiceProvider.notifier)
+                        .updateThemeMode(newTheme);
+                  },
+                  label: const Text('Theme'),
+                  dropdownMenuEntries: const [
+                    DropdownMenuEntry(
+                      value: ThemeMode.system,
+                      label: 'System Theme',
+                      leadingIcon: Icon(Icons.brightness_medium),
+                    ),
+                    DropdownMenuEntry(
+                      value: ThemeMode.light,
+                      label: 'Light Theme',
+                      leadingIcon: Icon(Icons.brightness_5),
+                    ),
+                    DropdownMenuEntry(
+                      value: ThemeMode.dark,
+                      label: 'Dark Theme',
+                      leadingIcon: Icon(Icons.brightness_3),
+                    ),
+                  ],
+                ),
               ),
             ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(16),
-            child: Container(
-              alignment: Alignment.topLeft,
-              child: FilledButton(
-                onPressed: () async {
-                  await ref.read(authServiceProvider.notifier).logOutUser();
-                  if (!context.mounted) return;
-                  await context.router.replace(LogInRoute());
-                },
-                style: ButtonStyle(
-                  shape: WidgetStateProperty.all(
-                    RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(5),
+            Padding(
+              padding: const EdgeInsets.all(16),
+              child: Container(
+                alignment: Alignment.topLeft,
+                child: FilledButton(
+                  onPressed: () async {
+                    await ref.read(authServiceProvider.notifier).logOutUser();
+                    if (!context.mounted) return;
+                    await context.router.replace(LogInRoute());
+                  },
+                  style: ButtonStyle(
+                    shape: WidgetStateProperty.all(
+                      RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(5),
+                      ),
                     ),
                   ),
+                  child: const Text('Log Out'),
                 ),
-                child: const Text('Log Out'),
               ),
             ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(16),
-            child: Container(
-              alignment: Alignment.topLeft,
-              child: ElevatedButton(
-                onPressed: () {
-                  showAboutDialog(
-                    context: context,
-                    applicationIcon: Assets.icons.logo.image(width: 50),
-                    applicationName: 'Town Talk',
-                    applicationVersion: packageVersion,
-                    applicationLegalese: '© 2024 Eli D. and Matthew W.',
-                    children: [
-                      const SizedBox(height: 24),
-                      Link(
-                        uri: Uri.https('github.com', 'PHS-TSA/nexus'),
-                        target: LinkTarget.blank,
-                        builder: (context, followLink) {
-                          return _AppDescription(followLink: followLink);
-                        },
-                      ),
-                    ],
-                  );
-                },
-                child: const Text('About'),
+            Padding(
+              padding: const EdgeInsets.all(16),
+              child: Container(
+                alignment: Alignment.topLeft,
+                child: ElevatedButton(
+                  onPressed: () {
+                    showAboutDialog(
+                      context: context,
+                      applicationIcon: Assets.icons.logo.image(width: 50),
+                      applicationName: 'Town Talk',
+                      applicationVersion: packageVersion,
+                      applicationLegalese: '© 2024 Eli D. and Matthew W.',
+                      children: [
+                        const SizedBox(height: 24),
+                        Link(
+                          uri: Uri.https('github.com', 'PHS-TSA/nexus'),
+                          target: LinkTarget.blank,
+                          builder: (context, followLink) {
+                            return _AppDescription(followLink: followLink);
+                          },
+                        ),
+                      ],
+                    );
+                  },
+                  child: const Text('About'),
+                ),
               ),
             ),
-          ),
-        ],
+          ],
+        ),
       ),
     );
   }
