+++ packages/app/lib/src/utils/responsive.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+/// Extension on [BuildContext] for responsive design.
+extension Responsive on BuildContext {
+  /// Returns true if the current screen size is mobileish.
+  bool get isMobile {
+    return MediaQuery.sizeOf(this).width <= 680;
+  }
+}
